+++ R/MsBackendPython.R
@@ -54,13 +54,20 @@
 #' - `backendInitialize()`: initializes the backend with information from the
 #'   referenced Python variable (attribute). The name of this attribute,
 #'   ideally stored in the associated Python session, is expected to be
-#'   provided with the `pythonVariableName` parameter. The optional
-#'   `spectraVariableMapping` parameter allows to provide additional, or
-#'   alternative, mapping of `Spectra`'s *spectra variables* to metadata in the
-#'   `matchms.Spectrum` objects. See [defaultSpectraVariableMapping()] (the
-#'   default) for more information. Parameter `pythonLibrary` must be used
-#'   to specify the Python library representing the MS data in Python. It can
-#'   be either `pythonLibrary = "matchms"` (the default) or
+#'   provided with the `pythonVariableName` parameter. This variable is
+#'   expected to contain or represent the full MS data. Alternatively, by
+#'   providing the full MS data with the `data` parameter, this MS data will
+#'   be first converted to the respective Python MS data representation and
+#'   the `MsBackendPy` object will be initialized as described before. This
+#'   allows to initialize the Python-based MS backend with data initially
+#'   available in R.
+#'   The optional `spectraVariableMapping` parameter allows to provide
+#'   additional, or alternative, mapping of `Spectra`'s *spectra variables*
+#'   to metadata in the `matchms.Spectrum` objects. See
+#'   [defaultSpectraVariableMapping()] (the default) for more information.
+#'   Parameter `pythonLibrary` must be used to specify the Python library
+#'   representing the MS data in Python. It can be either
+#'   `pythonLibrary = "matchms"` (the default) or
 #'   `pythonLibrary = "spectrum_utils"`. The function returns an initialized
 #'   instance of `MsBackendPy`.
 #'
@@ -100,7 +107,10 @@
 #'     names of the peaks variables to retrieve.
 #'
 #' @param data For `backendInitialize()`: `DataFrame` with the full MS data
-#'     (peaks data and spectra data). Currently not supported.
+#'     (peaks data and spectra data) such as extracted with the
+#'     [Spectra::spectraData()] method on another `MsBackend` instance.
+#'     Importantly, the `DataFrame` must have columns `"mz"` and
+#'     `"intensity"` with the full MS data.
 #'
 #' @param drop For `spectraData()` and `peaksData()`: `logical(1)` whether,
 #'     when a single column is requested, the data should be returned as a
@@ -162,11 +172,27 @@
 #' ## conversions.
 #' py_set_attr(py, "s_p", rspec_to_pyspec(s))
 #'
-#'
 #' ## Create a `MsBackendPy` representing an interface to the data in the
 #' ## "s_p" variable in Python:
 #' be <- backendInitialize(MsBackendPy(), "s_p")
 #' be
+#'
+#' ## An easier way to change the data representation of a `Spectra` object
+#' ## from R to Python is to use the `Spectra`'s `setBackend()` method
+#' ## selecting a `MsBackendPy` as the target backend representation:
+#' s_2 <- setBackend(s, MsBackendPy(), pythonVariableName = "s_p2")
+#' s_2
+#'
+#' ## This moved the data from R to Python, storing it in a Python variable
+#' ## with the name `s_p2`. The resulting `s_2` is thus a `Spectra` object
+#' ## with all MS data however stored in Python.
+#'
+#' ## Alternatively, by passing the full MS data with parameter `data`, the
+#' ## data is first converted to Python and the backend is initialized with
+#' ## that data. The `setBackend()` call from above internally uses this
+#' ## code to convert the data.
+#' be <- backendInitialize(MsBackendPy(), "s_p3",
+#'     data = spectraData(s, c(spectraVariables(s), "mz", "intensity")))
 #'
 #' ## Create a Spectra object which this backend:
 #' s_2 <- Spectra(be)
@@ -261,16 +287,15 @@ setMethod("backendInitialize", "MsBackendPy",
           function(object, pythonVariableName = character(),
                    spectraVariableMapping = defaultSpectraVariableMapping(),
                    pythonLibrary = c("matchms", "spectrum_utils"),
-                   ...,
-                   data) {
+                   ..., data) {
               .check_spectra_variable_mapping(spectraVariableMapping)
               object@spectraVariableMapping <- spectraVariableMapping
               pythonLibrary <- match.arg(pythonLibrary)
               if (!length(pythonVariableName))
-                  stop("'pythonVariableName' has to be provided")
+                  stop("'pythonVariableName' has to be provided", call. = FALSE)
               if (!is.character(pythonVariableName))
                   stop("'pythonVariableName' is expected to be the name ",
-                       "of the variable with the data.")
+                       "of the variable with the data.", call. = FALSE)
               if (missing(data)) {
                   ## 1) Check if the variable exists - and whether it's in py or
                   ##    in R. Set the `is_in_py` variable depending on that.
@@ -285,11 +310,26 @@ setMethod("backendInitialize", "MsBackendPy",
                   object@py_var <- pythonVariableName
                   object <- reindex(object)
                   object@py_lib <- pythonLibrary
+                  object
               } else {
-                  stop("Support for 'data' parameter is not yet implemented")
-                  ## if data provided -> convert the data to Python.
+                  if (!is(data, "DataFrame"))
+                      stop("'data' is expected to be a 'DataFrame'",
+                           call. = FALSE)
+                  if (!all(c("mz", "intensity") %in% colnames(data)))
+                      stop("Columns \"mz\" and \"intensity\" are required",
+                           call. = FALSE)
+                  py_set_attr(
+                      py, pythonVariableName,
+                      switch(pythonLibrary,
+                             matchms = .rspec_to_matchms_pyspec(
+                                 data, mapping = spectraVariableMapping),
+                             spectrum_utils = .rspec_to_spectrum_utils_pyspec(
+                                 data, mapping = spectraVariableMapping)))
+                  backendInitialize(
+                      object, pythonVariableName = pythonVariableName,
+                      spectraVariableMapping = spectraVariableMapping,
+                      pythonLibrary = pythonLibrary)
               }
-              object
           })
 
 #' @importMethodsFrom methods show